requirements.txt
torch>=2.1.0
transformers>=4.36.0
pettingzoo[mpe]>=1.24.0
gymnasium>=1.0.0
numpy>=1.24.0
pyyaml>=6.0
tensorboard>=2.14.0
wandb>=0.16.0
tqdm>=4.66.0
